change floorplan swap servie

+++ app/services/yardi2_swap_service.rb
@@ -38,9 +38,8 @@ class Yardi2SwapService < BaseService
               ils_units << pr[1]
             end
           end
-
-          save_yardi2_units(ils_units,external_property_id)
           save_yardi2_floorplans(floorplans)
+          save_yardi2_units(ils_units,external_property_id)
           end
       rescue => e
         e.message
@@ -147,7 +146,7 @@ class Yardi2SwapService < BaseService
       begin
 
 
-        fp = Floorplan.where(community_id: credentials.community_id,name: floorplan[0][:Name]).first
+        fp = Floorplan.where(community_id: credentials.community_id,name: floorplan[1][:Name]).first
         if fp.present?
           rooms = []
           floorplan.each do |f|
@@ -155,8 +154,11 @@ class Yardi2SwapService < BaseService
             if f.key?(:Room)
               rooms << f
             end
-            fp.provider_floorplan_id= f[:Id]
 
+            if f.key?(:Id)
+              fp.provider_floorplan_id= f[:Id]
+              fp.provider = "yardi_new"
+            end
             if f.key?(:MarketRent)
               if f[:MarketRent][0][:Min].to_f > 0
                 fp.market_rent = f[:MarketRent][0][:Min]
@@ -186,19 +188,25 @@ class Yardi2SwapService < BaseService
           fp.save(validate: false)
         else
 
+
+          fp = Floorplan.new
+          rooms = []
           floorplan.each do |f|
-            dup = Floorplan.find_by(community_id: credentials.community_id,provider_floorplan_id: f[:Id])
-            if dup.present?
-              dup.destroy
-            end
-            fp = Floorplan.new
+
             fp.community_id = credentials.community_id
-            rooms = []
-            fp.provider = "yardi_new"
+
+
             if f.key?(:Room)
               rooms << f
             end
-            fp.provider_floorplan_id= f[:Id]
+            if f.key?(:Id)
+              dup = Floorplan.find_by(community_id: credentials.community_id,provider_floorplan_id: floorplan[0][:Id])
+              if dup.present?
+                dup.destroy
+              end
+              fp.provider_floorplan_id = f[:Id]
+              fp.provider = "yardi_new"
+            end
             if f.key?(:Name)
               fp.name = f[:Name]
             end
@@ -228,7 +236,6 @@ class Yardi2SwapService < BaseService
               fp.bathrooms = room[:Room][1][:Count]
             end
           end
-
           fp.save(validate: false)
         end
